requirements.txt
dnspython~=2.2.1
dnslib~=0.9.19